Remove use of NSDecrementRefCountWasZero.
This function is unreliable and is impossible to implement safely in the
persence of ARC (even in other modules).  ARC requires that the runtime delete
all weak references before object deallocation and doesn't permit resurrection.

This change is initially motivated by the desire to steal a bit from the
refcount to indicate that an object has had weak references taken.