Refactor weak symbol usage.
The Linux run-time linker doesn't allow weak references in one library
to be overridden by ones in another.  To work around this, we now
declare the runtime functions as weak and perform dynamic checks on
whether the symbols have been resolved and call the fallbacks as local
static functions if they are not present.