Fix GSICUString cloning.
We were previously not copying the state, nor were we making sure that
the buffers pointed to the correct places.  This didn't matter for how
GSICUString was used in GNUstep, but it was problematic for other uses.

Also correctly handle the case that we can't allocate extra buffer
space.  We should allow the user to specify the correct thing.