Add a .clang-format file
It gets hard to code in GNU style for people with editors that don't
understand the difference between indentwidth and tabstop.

This file is based on gcc's contrib/clang-format. A flaw of it is that
it does not put a space after selector colons. I am gonna file a LLVM
bug for that.