Separate __has_attribute test from actual usage, as per GCC documentaiton.
The first `#if' test succeeds only when the operator is supported by the version of GCC (or another compiler) being used. Only when that test succeeds is it valid to use __has_attribute as a preprocessor operator. As a result, combining the two tests into a single expression as shown below would only be valid with a compiler that supports the operator but not with others that don't.

https://gcc.gnu.org/onlinedocs/cpp/_005f_005fhas_005fattribute.html