NSBundle: fix es.lproj path not generated
The conversion between BCP 47 language tags and
the old-style language names ('English', 'German', etc) is quite
complex, mapping ISO 639-1 language tags to the old-style names, and
the result back to ISO 639-1 (-ish).

The ISO 639-1 tag for spanish 'es' maps to 'Spanish', which in turn
maps back to 'sp'. The original input is not added to the list
of possible names for 'lproj' folders.

We fix this by operating on the official input, when splitting
language, variant, and region. The "mapped" tag is always added
just to keep 'sp.lproj' working.