NSTimeZone: Use GetDynamicTimeZoneInformation
tz.StandardName is not always available, does not represent the raw
timezone key from the registry. It appears that ucal_getTimeZoneIDForWindowsID
maps the registry key to a compatible IANA time zone.

The dotnet and flutter implementations also use
GetDynamicTimeZoneInformation and use TimeZoneKeyName.

Here is an example for Asia/Singapore:
StandardTime = Malay Peninsula Standard Time
TimeZoneKeyName = Singapore Standard Time

StandardTime is not registered in the IANA database.