NSProgress: Reimplement according to specification
The current implementation of NSProgress in GNUstep is outright wrong,
and is not implemented according to the public API documentation
provided by Apple.

While Apple's documentation omits important implementation details,
even explicitly mentioned behaviour was either not implemented or
invented.

This commit provides a complete reimplementation of the NSProgress
API. The following process was used to develop an implementation
close to the one in Foundation:

1. Write unit tests based on the description in the Apple API
   documentation.
2. Validate the tests on macOS 15.5
3. Write more unit tests of implementation-specific and undocumented
   behaviour (including order of KVO event messages).
4. Implementation in GNUstep